Fix: Keras 3.x, yfinance 1.x e qualidade de código em src/
- model.py / predict.py: substitui tensorflow.keras por import keras direto (Keras 3.x)
- api/main.py: remove rota raiz duplicada que conflitava com routes.py
- api/routes.py: adiciona decorator @router.get('/') em _build_dashboard; corrige raise-from em todos os except
- evaluate.py: remove double-cast float() redundante em rmse e mape
- train.py: substitui datetime.utcnow() por datetime.now(timezone.utc) (Python 3.12+)
- data_loader.py: remove sessão curl_cffi incompatível com yfinance 1.x; usa tz_convert ao invés de tz_localize

+++ src/evaluate.py
@@ -26,7 +26,7 @@ def compute_metrics(y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
     y_pred = np.asarray(y_pred).flatten()
 
     mae = mean_absolute_error(y_true, y_pred)
-    rmse = float(np.sqrt(mean_squared_error(y_true, y_pred)))
+    rmse = np.sqrt(mean_squared_error(y_true, y_pred))
 
     # MAPE manual para evitar divisão por zero
     mask = y_true != 0
@@ -35,7 +35,7 @@ def compute_metrics(y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
     return {
         "mae": float(mae),
         "rmse": float(rmse),
-        "mape": float(mape),
+        "mape": mape,
     }
